Adding linux makefile to git.
It's pretty annoying to have to compile the Projucer in GPL mode
in order to Build for linux. While I've been depending on the
Projucer for generating IDE projects (as opposed to checking
IDE projects into git), because otherwise you have commit the
IDE project AND cybr.jucer every time you want to add a cpp/h file
to avoid accidentally checking in an out-of-date project file.

However, for now I think it will be much easier to commit the Makefile
so that its easier to compile the projecer just for the sake of
compiling cybr on linux.